Fix a warning.

+++ MP42/MP42SubUtilities.m
@@ -136,7 +136,7 @@ canOutput:
 
 -(SBSubLine*)getSerializedPacket
 {
-	int nlines = [lines count];
+	NSUInteger nlines = [lines count];
     
 	if (!nlines) return nil;
